Finish method for inserting random beers in DB. Add beer.properties & config for it. Change visibilities for some properties in PunkApi*Dtos.

+++ src/main/java/com/example/dibchallenge/beer/strategy/BeerStrategy.java
@@ -6,6 +6,6 @@ public interface BeerStrategy {
 
     BeerModel getRandomBeer();
 
-    List<BeerModel> getRandomBeers(int count);
+    List<BeerModel> getRandomBeers(long count);
 
 }
